requirements.txt
numpy>=1.24
pandas>=2.0
plotly>=5.22
matplotlib>=3.4
selenium>=4.21
Pillow>=10.0
